+++ phoenix-core/src/main/java/com/salesforce/phoenix/query/ConnectionlessQueryServicesImpl.java
@@ -274,7 +274,7 @@ public class ConnectionlessQueryServicesImpl extends DelegateQueryServices imple
     }
     
     @Override
-    public List<TableName> reserveSequences(String tenantId, Set<Map.Entry<TableName,SequenceValue>> sequences, long batchSize) throws SQLException {
+    public List<TableName> reserveSequences(String tenantId, Set<Map.Entry<TableName,SequenceValue>> sequences, int batchSize) throws SQLException {
         return Collections.emptyList();
     }
